Python_server/requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pycparser==2.22
PyMuPDF==1.25.1
pypdfium2==4.30.0
pytesseract==0.3.13
Werkzeug==3.1.3
